components/train_test_split/requirements.txt
pandas
mlflow
scikit-sklearn